pyproject.toml
[project]
name = "nhl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.3",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
]
